Improved matching of multi-word "cussWords"
Entries in the cussWords array which contained spaces were not matched when detectEvasionCharacters was set to true. By first running the cussword through a regex to remove spaces, and then replacing the de-spaced match in a string with the version from the cussword array, those entries are now correctly detected.